// frontend/src/enums/KeelType.ts
export enum KeelType {
    ALL_KEEL_TYPES = "All keel types",
    BILGE_KEEL = "Bilge/Twin Keel",
    BULB_KEEL = "Bulb Keel",
    CANTING_KEEL = "Canting Keel",
    CENTREBOARD = "Centreboard & Daggerboard",
    FIN_KEEL = "Fin Keel",
    FULL_KEEL = "Full Keel",
    LIFTING_KEEL = "Lifting & Swing Keel",
    LONG_KEEL = "Long Keel",
    SHOAL_KEEL = "Shoal Keel",
    WINGED_KEEL = "Winged Keel"
}